Fix reg_read_batch(), test registers
This commit made a backwards-incompatible change; instead of accepting a table of register IDs
to read and returning a table of values, reg_read_batch() now accepts a variable number of
arguments and returns a variable number of results. This is easier to implement and also allows
the caller to collect the results individually or as a single table.